pyproject.toml
[project]
name = "dss-irrigation"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.1.0",
    "kaleido>=1.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-fuzzy>=0.5.0",
    "scikit-learn>=1.8.0",
    "tensorboard>=2.20.0",
    "torch>=2.10.0",
    "xgboost>=3.1.3",
]
